mikroserwis_eeg/brainaccess_board/pyproject.toml
[project]
name = "brainaccess-board"
version = "1.0.0"
description = "Python library for interfacing with BrainAccess Board"
dependencies = [
    "appdirs>=1.4.4",
    "pydantic>=2.7.3",
    "numpy>=1.26.4",
    "pyzmq>=26.0.3",
    "mne>=1.8.0",
    "pylsl>=1.16.2",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["brainaccess_board"]
